Move InventoryFile class to sphinx.util.inventory
